Joel Oliveira: More explicit method signature
For other developers coming into this class there may be some negligible benefits for passing class method params on to instance methods with a splat if the method signatures are identical and not  comprised of parameters that could change based on the content.

In other words - "don't make me work more than I need to" ;)

This commit is, essentially, a request to make those class method parameters more explicit and offload the cognitive work of tracking down the method signature(s) of the instance method.