Joel Oliveira: Split encoded value and name. Add specs.
Based on the documentation in the [foxycart wiki] there is some differing behavior based on when the value and name are being assigned the encoded HMAC data. Specifically:

* When there is a numeric prefix to the form name the encoded data should not use `<number>:`.
* When the value is an OPEN value, the result should contain "||open" appended to the end.
* Etc ...

To keep track of these cases I've added some tests for both "CYA" test coverage, and documentation purposes. It's easier to reason the edge cases when we can hold up the example inputs against the results.

[foxycart wiki]: https://wiki.foxycart.com/v/2.0/hmac_validation#an_example