Use class object instead of string for middleware
When upgrading to Rails 5.1.4 I ran into an issue where adding the
FoxycartHelpers middleware resulted in an exception. This commit changes
the way we insert the FoxycartHelpers middleware - with a class object
instead of the string representation of the class.

Also - explicitly require OpenSSL in order for the test suite to get
back to passing!